refactor(quota): remove quota rule cache (#322)
