+++ packages/core/src/index.ts
@@ -27,4 +27,3 @@ export * from './watermark/watermark'
 export * from './gotenberg/gotenberg'
 export * from './mcp/mcp-service'
 export * from './quota/quota-service'
-export * from './quota/quota-cache'
